Разделение программ
дабы не искать каждый раз что нужно менять, программы разделились
